Notiva/pyproject.toml
[project]
name = "notiva"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "manas-099", email = "manaspatra481@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.111.0",
    "langchain-openrouter>=0.1.0",
    "numpy>=2.4.3",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "rich>=14.3.3",
    "sarvamai>=0.1.0",
    "sendgrid>=6.11.0",
    "sounddevice>=0.4.6",
    "structlog>=24.1.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=16.0",
]

[project.scripts]
notiva = "notiva:main"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"
